refactor: move layout shell to app/[locale]/components and update all importers
