+++ app/app/[locale]/components/entry_options/index.tsx
