+++ app/app/[locale]/components/header/index.tsx
@@ -4,7 +4,7 @@ import { Prop } from './types';
 import { TAppendClass } from 'components/button/types';
 import { useState } from 'react';
 import { useTranslations } from 'next-intl';
-import { Link } from '../../i18n/navigation';
+import { Link } from 'i18n/navigation';
 import Image from 'next/image';
 import { Button } from 'components';
