+++ app/app/[locale]/components/header/types/index.ts
