+++ app/app/[locale]/components/layout/types/index.ts
