+++ app/app/[locale]/components/popups/types/index.ts
